move color and type variations to different locations in the UI
