+++ packages/edit-site/src/components/global-styles/variations-typography.js
@@ -1,0 +1,266 @@
+/**
+ * External dependencies
+ */
+import classnames from 'classnames';
+
+/**
+ * WordPress dependencies
+ */
+import { store as coreStore } from '@wordpress/core-data';
+import { useSelect } from '@wordpress/data';
+import { useMemo, useContext, useState } from '@wordpress/element';
+import { ENTER } from '@wordpress/keycodes';
+import {
+	__experimentalHeading as Heading,
+	__experimentalGrid as Grid,
+	__experimentalVStack as VStack,
+} from '@wordpress/components';
+import { __, sprintf } from '@wordpress/i18n';
+import { privateApis as blockEditorPrivateApis } from '@wordpress/block-editor';
+
+/**
+ * Internal dependencies
+ */
+import { mergeBaseAndUserConfigs } from './global-styles-provider';
+import { unlock } from '../../lock-unlock';
+import { getFamilyPreviewStyle } from './font-library-modal/utils/preview-styles';
+
+function cloneDeep( object ) {
+	return ! object ? {} : JSON.parse( JSON.stringify( object ) );
+}
+
+const filterObjectByProperty = ( object, property ) => {
+	const newObject = {};
+	Object.keys( object ).forEach( ( key ) => {
+		if ( key === property ) {
+			newObject[ key ] = object[ key ];
+		} else if ( typeof object[ key ] === 'object' ) {
+			const newFilter = filterObjectByProperty( object[ key ], property );
+			if ( Object.keys( newFilter ).length ) {
+				newObject[ key ] = newFilter;
+			}
+		}
+	} );
+	return newObject;
+};
+
+const removePropertyFromObject = ( object, property ) => {
+	for ( const key in object ) {
+		if ( key === property ) {
+			delete object[ key ];
+		} else if ( typeof object[ key ] === 'object' ) {
+			removePropertyFromObject( object[ key ], property );
+		}
+	}
+	return object;
+};
+
+const getVariationsByType = ( user, variations, type ) => {
+	const userSettingsWithoutType = removePropertyFromObject(
+		cloneDeep( user ),
+		type
+	);
+
+	const variationsWithOnlyType = variations.map( ( variation ) => {
+		return filterObjectByProperty( variation, type );
+	} );
+
+	return variationsWithOnlyType.map( ( variation ) =>
+		mergeBaseAndUserConfigs( userSettingsWithoutType, variation )
+	);
+};
+
+const { GlobalStylesContext, areGlobalStyleConfigsEqual } = unlock(
+	blockEditorPrivateApis
+);
+
+const getFontFamilies = ( themeJson ) => {
+	const headingFontFamilyCSS =
+		themeJson?.styles?.elements?.heading?.typography?.fontFamily;
+	const headingFontFamilyVariable =
+		headingFontFamilyCSS &&
+		headingFontFamilyCSS.replace( 'var(', '' ).replace( ')', '' );
+	const headingFontFamilySlug = headingFontFamilyVariable
+		?.split( '--' )
+		.slice( -1 )[ 0 ];
+
+	const bodyFontFamilyVariable = themeJson?.styles?.typography?.fontFamily
+		.replace( 'var(', '' )
+		.replace( ')', '' );
+
+	const bodyFontFamilySlug = bodyFontFamilyVariable
+		?.split( '--' )
+		.slice( -1 )[ 0 ];
+
+	const fontFamilies = themeJson?.settings?.typography?.fontFamilies?.theme; // TODO this could not be under theme.
+
+	const bodyFontFamily = fontFamilies.find(
+		( fontFamily ) => fontFamily.slug === bodyFontFamilySlug
+	);
+
+	let headingFontFamily = fontFamilies.find(
+		( fontFamily ) => fontFamily.slug === headingFontFamilySlug
+	);
+
+	if ( ! headingFontFamily ) {
+		headingFontFamily = bodyFontFamily;
+	}
+
+	return [ bodyFontFamily, headingFontFamily ];
+};
+const getFontFamilyNames = ( themeJson ) => {
+	const [ bodyFontFamily, headingFontFamily ] = getFontFamilies( themeJson );
+	return [ bodyFontFamily?.name, headingFontFamily?.name ];
+};
+
+const normalizedHeight = 100;
+const ratio = 1;
+
+function TypographyVariation( { variation } ) {
+	const [ isFocused, setIsFocused ] = useState( false );
+	const { base, user, setUserConfig } = useContext( GlobalStylesContext );
+	const context = useMemo( () => {
+		return {
+			user: {
+				settings: variation.settings ?? {},
+				styles: variation.styles ?? {},
+			},
+			base,
+			merged: mergeBaseAndUserConfigs( base, variation ),
+			setUserConfig: () => {},
+		};
+	}, [ variation, base ] );
+
+	const selectVariation = () => {
+		setUserConfig( () => {
+			return {
+				settings: variation.settings,
+				styles: variation.styles,
+			};
+		} );
+	};
+
+	const selectOnEnter = ( event ) => {
+		if ( event.keyCode === ENTER ) {
+			event.preventDefault();
+			selectVariation();
+		}
+	};
+
+	const isActive = useMemo( () => {
+		return areGlobalStyleConfigsEqual( user, variation );
+	}, [ user, variation ] );
+
+	let label = variation?.title;
+	if ( variation?.description ) {
+		label = sprintf(
+			/* translators: %1$s: variation title. %2$s variation description. */
+			__( '%1$s (%2$s)' ),
+			variation?.title,
+			variation?.description
+		);
+	}
+
+	const [ bodyFontFamilies, headingFontFamilies ] = getFontFamilies(
+		mergeBaseAndUserConfigs( base, variation )
+	);
+
+	const bodyPreviewStyle = getFamilyPreviewStyle( bodyFontFamilies );
+	const headingPreviewStyle = {
+		...getFamilyPreviewStyle( headingFontFamilies ),
+		fontSize: '1.2rem',
+	};
+
+	return (
+		<GlobalStylesContext.Provider value={ context }>
+			<div
+				className={ classnames(
+					'edit-site-global-styles-variations_item',
+					{
+						'is-active': isActive,
+					}
+				) }
+				role="button"
+				onClick={ selectVariation }
+				onKeyDown={ selectOnEnter }
+				tabIndex="0"
+				aria-label={ label }
+				aria-current={ isActive }
+				onFocus={ () => setIsFocused( true ) }
+				onBlur={ () => setIsFocused( false ) }
+			>
+				<VStack
+					className="edit-site-global-styles-variations_item-preview"
+					isFocused={ isFocused }
+					style={ {
+						height: normalizedHeight * ratio,
+						lineHeight: 1.2,
+						textAlign: 'center',
+					} }
+				>
+					<div style={ headingPreviewStyle }>
+						{ headingFontFamilies.name }
+					</div>
+					<div style={ bodyPreviewStyle }>
+						{ bodyFontFamilies.name }
+					</div>
+				</VStack>
+			</div>
+		</GlobalStylesContext.Provider>
+	);
+}
+
+export default function TypographyVariations() {
+	const variations = useSelect( ( select ) => {
+		return select(
+			coreStore
+		).__experimentalGetCurrentThemeGlobalStylesVariations();
+	}, [] );
+
+	const { base, user } = useContext( GlobalStylesContext );
+
+	const typographyVariations =
+		variations && getVariationsByType( user, variations, 'typography' );
+
+	const uniqueTypographyVariations = [];
+	const uniqueTypographyNames = [];
+	const isDup = ( x, y ) => {
+		return uniqueTypographyNames.find( ( it ) => {
+			return JSON.stringify( it ) === JSON.stringify( [ x, y ] );
+		} );
+	};
+
+	typographyVariations?.forEach( ( variation ) => {
+		const [ bodyFontFamilyName, headingFontFamilyName ] =
+			getFontFamilyNames( mergeBaseAndUserConfigs( base, variation ) );
+		if ( ! isDup( bodyFontFamilyName, headingFontFamilyName ) ) {
+			uniqueTypographyVariations.push( variation );
+			uniqueTypographyNames.push( [
+				bodyFontFamilyName,
+				headingFontFamilyName,
+			] );
+		}
+	} );
+
+	return (
+		<>
+			<div className="edit-site-sidebar-navigation-screen-styles__group-header">
+				<Heading level={ 3 }>{ __( 'Typography' ) }</Heading>
+			</div>
+			<Grid
+				columns={ 2 }
+				className="edit-site-global-styles-style-variations-container"
+			>
+				{ uniqueTypographyVariations &&
+					uniqueTypographyVariations.map( ( variation, index ) => {
+						return (
+							<TypographyVariation
+								key={ index }
+								variation={ variation }
+							/>
+						);
+					} ) }
+			</Grid>
+		</>
+	);
+}
